Update sort tests to reflect new info object and new stochastic sort
Most of these changes are to make sure the redirectInfo object is picking
up the values it should.

There's also a new test called "test-adaptive-sort" that covers a
previously-untested set of code. When I started writing the test, I found
the infinite loop bug mentioned in this PR's previous commit.
NOTE: The new test is not super rigorous because I don't know that it makes
a lot of sense to spend time designing a good test for something as stochastic
as our adaptive sort. In the future, when we're convinced adaptive sort does
exactly what we want, we should revisit the test and make sure its assumptions
hold true.