Plumb new multi-issuer logic into Scitokens config and Director ads
This handles creation of the correct scitokens config for the Origin,
and sends the relevant information to the Director. That I can tell,
both from unit tests and a few hand-crafted checks, the information that
arrives in the Director is sufficient for caches to turn around and
construct their own correct scitokens config for multi-issuer Origins.