Pre-allocate available ports for xrootd tests
Because issuer configuration now happens on instantiation of the Origin
exports, we really do need to know the ports ahead of time. Unfortunately
this breaks some of our testing infrastructure, which relied on a feature
in XRootD that turned a port of '0' into a random, available port handed
over by the OS.

However, setting '0' as the port causes many of the default issuers to use
this port, both for the issuer proper and the server's audience URL in
scitokens config. Manually correcting ports in this file would be complicated
and would require waiting for XRootD to sit around long enough that it decides
to re-load the file.

Instead, this pre-allocates the ports by asking the OS for a few that are
available. There's a small race condition in testing environments where
the ports we get back from the OS might become bound before our Pelican
servers start. Hopefully this doesn't make testing super flaky, but I'm
assuming the odds of that happening are currently relatively minimal.