Switch 'Server.DirectorURLs' to 'Server.DirectorUrls' for naming convention
I didn't notice on my first review of this code, but the capital URL breaks the
naming convention we use for all the other "Url"-bearing params. Since this isn't
merged at the time of this commit, there is no backwards compat issue with changing
it.

Another good reason to use 'param.XXX.GetName()' in log messages is that updating
a config param in 'docs/parameters.yaml' forces us to fix the log messages. Luckily
I noticed a log message that hardcoded the old 'DirectorURLs'.