Adapt the test cases to the new ResetTestState()
The new ResetTestState() has already incorporated the serverAds reset. So the code for the reset of the serverAds is not needed in the test cases. Also ResetTestState() just needs to be called at the beginning and the end of a test function, otherwise it will reset the serverAds between each sub-test, leading to the sub-test failure.