Introduce feature compatibility package
This commit introduces a framework for determining whether a particular
server is compatible with a "feature" the server may or may not support.
In trying to keep this framework flexible, I decided that each feature
itself may have a version, and that those versions may depend on specific
versions of each of the other Pelican servers we operate (cache, origin,
director, registry). Eventually we should provide an override for the
compatibility yaml so server admins can make on-the-fly corrections.

This diff also introduces a ternary type in our utils package.
There have been several places in Pelican where it'd be useful to
represent ternary decisions where valid states can be "true", "false",
or "unknown". Another recent example was the PR introducing high
availability Directors. While that still uses its own ternary representation
(wanted to minimize scope creep so I didn't touch it), we should eventually
land on a common base for this type of logic.

Ternary is needed in feature compatibility so that we don't need to enumerate
every possible compatibility between versions and features.