Add notion of required features to origin ads
This adds the names of required features to the origin's advertisement
struct, which can then be used by the Director during matchmaking. We
might eventually want to implement the parallel notion of required
features in caches, but there's no reason to do that yet.

When these feature names arrive at the Director, a helper function converts
them back into the feature struct using the helper function in the new
features library.