Overhaul Director matchmaking so we can filter based on feature compat
This commit (quite a big one) overhauls some of the Director's matchmaking
and server ad sorting so that we can more easily reason over a variety of
matchmaking policies. In the scope of the encapsulating PR, it's then used
to filter caches based on whether they support an origin's CacheAuthz
requirements.

As for the design of the filtering mechanism, each filter function uses
a list of predicate filter funcs so that we can iterate through a slice
of ads exactly once and weed out any ad that fails one of the predicates.

In addition to the overhaul, I moved a few core functions around to better
reflect their purposes. For example, I moved `getAdsForPath` into `director/
sort.go` because this function is used in sorting.

Some features I touched, like cache chaining, have no real unit tests and
have never been turned on in production. It remains to be seen whether they
functioned properly in the first place, let alone whether I preserved that
functionality.