Sanity check received size and on-disk size for client GETs
When the client requests an object, grab determines the expected download
size by inspecting the `Content-Length` header. Previously we didn't perform
any checks against this value to make sure the number of bytes we receive
actually match the number of expected bytes -- we only used status codes
to determine success.

With this diff, we do two sanity checks. The first inspects the number of bytes
that ran through the grab client's writer, and the second actually looks at
on-disk metadata to make sure the correct number of bytes were persisted.

If the grab sanity check passes but there's an error with the stat, I don't
treat that as an error. I could easily be convinced that it _should_ be an
error, but I don't know enough about how stat is supposed to perform on all
systems to make that call at this point.