Fix nuanced Director stat bug and assign request UUIDs
This commit fixes a bug in the director that caused some origin requests
to stat all caches it might be aware of. I noticed this after seeing what
should've been a simple request generating many more cache requests than
it should've.

In debugging this, I found it very useful to track specific requests through
the Director, for which I added UUIDs to all the requests. Since I'm likely
to need this after this overhaul hits production, I cleaned it up and left
it in rather than removing it. Sorry for the scope creep...