Don't require specific error message for test whose failure mode isn't being tested
This test started failing after a subtle change in stat behavior (which
happened after fixing a bug). The test's goal is simply to demonstrate
we catch a failure if trying to recursively download a non-existant dir,
but an assertion forced a specific error message.

This change gets rid of the assertion for a specific error message and
instead recognizes a failure as a failure.