Fix cache self-test error message & prevent file name truncation
As #1808 states, we'd observed something strange in the relevant functions
where a) the file URL generation function could return an empty URL without
an associated error and b) that URL file could be successfully cleaned up,
but logged in a way where the URL is empty.

In addition to tidying up a few other error messages, I believe this makes
it much harder for the conditions that caused this behavior to exist.