Flush logs in some error cases during InitServer()
There's a boostrapping issue here -- we don't generally want to flush logs
until we've ingested the web config, which may set a logging file.
We can't do that directly in `InitConfig()` because it requries a default
value that's only generated in `SetServerDefaults`, which is only run in
`InitServer()`. Thus, we have to manually flush some of these logs for certain
error cases that happen before we've ingested the web config.