Fix ability to set config dir with env var and isolate additional test
I'd been wondering for awhile why setting `PELICAN_CONFIGDIR` in tests
wasn't correctly isolating the tests. Like a bad developer, I worked
around this by reconfiguring those tests to use a specific config file
instead of digging into the issue.

Turns out the issue was due to mis-sequencing in `InitConfig()`. This
corrects the sequence so viper knows to look for `PELICAN_CONFIGDIR`
by the time it sets the config directory.

I also added extra isolation to a test that pointed me to the issue --
it was failing locally because it was picking up my system config in
/etc/pelican. Naughty!