Refactor HTTP downloads to include checksums
This admittedly-large commit (the history got rebased to death) adds
support for checksum verification in the HTTP downloads.  After a
download is complete, the client will do a second HEAD request (which
is permitted to fail; depends on the version of the underlying server)
requesting a checksum.  If a checksum is returned and it doesn't match
the client's calculated checksum, then the transfer is forced to fail.

This is accomplished by providing a new Writer overlay that uses Go's
built-in hash classes.  Since all download paths now override the default
grab client with a Writer, most of the reason why we were using grab
has now gone.  I took the opportunity to refactor it away; there's
now one less unmaintained dependency in Pelican!

The checksum writers are calculated outside the `downloadHttp` function,
meaning that if a transfer is resumed from a second location, then the
checksum will be maintained.

Because the project name is now retrieved much more frequently, instead
of re-parsing a file for each transfer, we parse it once and cache the
results globally..

There are still tests to write - but existing tests currently pass so
it's time to see how far we get on GitHub.