Pelicanize "NEAREST_CACHE" env var with 'Client.PreferredCaches'
This standardizes our use of `Client.PreferredCaches` internally as a
string slice, bearing the "+ means continue" semantics. Previously
we used an entirely undocumented environment variable `NEAREST_CACHE`,
and in only a subset of places we also checked for `PELICAN_NEAREST_CACHE`.
With the new approach, we still handle these configurations, but immediately
stuff them into viper to be used later in a standardized way.