Normalize rc versions when determining feature compatibility
Previously we would've run into an issue where a feature might be configured
to work starting with v7.16.0, but the version library treats v7.16.0-rc.0 as
"less than" that. Prerelease do after all happen before releases!

This would prevent us from testing whether a v7.16.0-rc.0 cache works with
CacheAuthz because the ServerSupportsFeature() function would treat the cache
as older than v7.16.0, which is when the feature is configured to become available.

Instead, this function has been changed to ignore release candidate metadata
and treat `v7.16.0-rc.0` the same as `v7.16.0`.