Don't refresh expiration on serverAds.Get()
The bug here was nuanced -- if a server had previously succeeded at
advertising but was restarted quickly in a way that caused advertisement
failure, the IO Stats routine would refresh the ad every 15 seconds,
completely preventing expiration. This adds the disable on touch
option so that these ads can expire.