Disable touch on hit for serverAds cache instantiation
I waffled a bit on whether to add this to the entire cache itself,
and ultimately decided I should. My reasoning is that currently,
we want to update expirations iff we're recording a new ad. In this
mindset, we never want a mere `serverAds.Get()` to refresh the exp.

I'm still leaving the other options for explicit `serverAds.Get()`
calls because my previous reasoning may become untrue someday, and
I want to future proof this against someone purposefully/accidentally
removing the option from the cache instantiation.