Use registration name and not hostname when requesting fed tokens
Previously the code always used the hostname. After we saw errors popping
up in production related to failed advertise token verification, I realized
we need the registration name, which is always the hostname for Origins,
and _sometimes_ the hostname for Caches. When `Xrootd.Sitename` is set in
caches, that's the value we should use.