Move 'identity' from QueryMyPrometheus sig into function body
It ocurred to me that this was another value we should move into the
function body for a few reasons...

It's not really valid to pass values like "director/Director",
"Cache", or "Origin" through the function signature. The JWT RFC states
that token subjects MUST either be unique within their issuer scope or
globally unique, so lumping all "Origin"s and "Director"s into a single
string is a violation.

Moreover, the function name implies we already know both the sender and
receiver of the token -- "me".

We now simply grabs the server's external Web URL inside the function
and use that to satisfy both subject uniqueness and standardize on
the values that are used.