Use better name for issuer matching function in cmd/token.go
The function in question is one that checks whether a given issuer URL
provides a public key whose KID field matches that of a given private key.

I decided not to use "verify" in the name, because that word is loaded
and often connotes some kind of cryptographic operation. Instead, this
function only aims to make sure the user generating a token with the
new `pelican token create` function doesn't use an issuer that couldn't
possibly work.