Validate configured hostnames against TLS certificate chain
This is a quality-of-life improvement that tries to help admins starting
a new Pelican server catch when their provided TLS cert files do not
match the configured hostnames of the new server.

Overall the validation works by checking different places a Pelican admin
may configure a hostname and then comparing that hostname against the
SANs and CN of the provided certificate.