Add 'Server.ExternalWebUrl' to self-signed certs to restore unit tests
A variety of unit tests may mix/match configurations of either
'Server.Hostname' and 'Server.ExternalWebUrl'. However, when only
'Server.ExternalWebUrl' is set in the tests, 'Server.Hostname' is
still always set based on the output of 'os.Hostname', meaning the
certificate would fail verification against the external URL.

Since these certificates are used primarily for internal unit testing,
it seemed that the correct solution to this was to add both hostnames
to the DNS names of the cert when needed.

While this fixed tests for me locally, it remains to be seen whether
I should also be adding the 'Cache.Url' or 'Origin.Url' hostnames to
the certificate. For now, I'll leave them out until I can justify the
need to include them.