Add some context errors to final error message
This vein of work tries to better expose the reason `WaitUntilWorking`
might be failing by attaching the last encountered error to the error
message we produce instead of leaving it detached in a log statement.

Previously I only attached errors that weren't context errors because
I hadn't seen a context error that wasn't a generic timeout. However,
while playing around with some of the tests, I discovered some context
errors were more useful -- such as one that indicated the TCP connection
was successful, but that the context timed out awaiting response headers.

To capture these without overwriting more useful errors by the final
context timeout that inevitably happens when the server doesn't pass
this probe, I now grab the context error IFF it is the first error.

In my testing, this captures the header timeout correctly.