Adjust `WaitUntilWorking` tests to work with modified error messages
I really dislike tests that rely on the particular wording of an error
message to pass, but this is one of those cases where it really doesn't
feel correct to spin off a whole bunch of typed errors only for easier
test validation.

To somewhat ameliorate the situation, I modified the existing tests
so that they're less dependent on an exact error message and instead
check for key words that I think are more likely to exist in any error
message regardless of the exact wording.

For example, if WaitUntilWorking expects an HTTP 200 but gets a 500,
I check that the error contains "expected 200" and "500".