Add IPs to self-baked TLS certs as IPs, not DNS names
Some tests use IP addresses from http test servers for various
Pelican server configs (e.g. 'Server.ExternalWebUrl'). When we do
this, the certificates Pelican generates for itself are invalid because
they contain a SAN that says the IP is a DNS entry.

This was exposed in tests for #2035 after it picked up the new
verification code that checks the cert against configured hostnames.