Use ternary results in 'doRequest' helper instead of sentinel error
Requested by the reviewer, this uses a ternary result instead of reyling
on a sentinel error to signal "success", "retry" and "fail".

I'd initially written this with a custom enum for the probe, but decided
it was basically identical to the ternary utility I wrote awhile ago so
I decided to use that instead with the caveat that 'Tern_Unknown' is
being used here to mean "retry". I believe this should be clear with the
provided comments.