Implement status-aware server sorting in the Director
This new section of code uses servers' reported status as part of the
sorting algorithm when the Director enables "adaptive" sort.

It works by starting with a raw weight 'x' in (0,1] that corresponds
to different statuses. From here, x at time t (xt) is adjusted to
take into account past known weights using an EWMA low-pass filter,
resulting in st. This smoothing function is applied to avoid sharp
box-wave-like changes in the server's priority due to status changes.

Effectively, this penalizes a server's priority proportional to the
length of time the server spends outside of the "ok" state. Perhaps
the most useful non-ok status to think about is the "degraded" status,
which occurs for servers that exceed some threshold relative to their
configured IO throttle limit. This means that servers under heavy load
(as defined by percentage of configured allowable connections) will
be slowly deprioritized for client redirects until the server has a
chance to dip below the configured limit.