Don't mod global viper in functions that take viper reference
There was another subtle bug here that was causing some things
to print warnings about defaults that should have otherwise been
set correctly. The bug was that we were mixing which viper instance
we operated on in the default setters for clients and servers.

Those functions take a viper reference, so we should only use the passed
viper for default setting logic. This means no `viper.Get` or
`param.<param name>.Get*()`, which operate on the global viper in
those functions.