Use namespace prefix token issuer as X-Pelican-Token-Generation issuer
Looking through various comments and past bugfixes, I'm really unsure
what the guiding philosophy surrounding this header is supposed to be,
and how that philosophy differs from the X-Pelican-Issuer header.

At a high level, there are issuers for a namespace and issuers for
the Origin server itself, but we've never formally explained how
these interact, let alone the configuration mechanisms needed to
cleanly split these as concepts. In any case, it seems like the token
generation header should be using an issuer URL tied to the namespace
itself.

This commit uses the first specified namespace issuer as the default
issuer for clients to bootstrap tokens.