Don't redirect to caches that report they're in the error state
Toward closing #2544, this commit ads another common predicate --
"CacheNotInErrorState" -- that caches must pass to be considered
for redirection.

The intent here is that whenever a cache self reports a "critical"
or worse condition, we shouldn't send clients to it. Such conditions
may be triggered by subsystem failures (e.g. failed Director tests),
and while they don't guarantee the cache couldn't handle the request,
the assumption is that these caches are much more likely to produce
client errors.

We don't treat Origins the same, since we often assume an Origin may
be the only source for an object. In this case, it's better to try
even if there's a high chance of errors.