Rework generation of client/server coordinates
The key idea in this commit is a re-think for how we derive and store
different types of coordinates.

Now, server and client coordinates are treated differently because
they have different error modes:
- An error is generated when we try to fetch a server coordinate and
  none can be resolved via override or maxmind
- No error is generated when we try to fetch a client coordinate because
  we're guaranteed to get one via random assignment as a last resort

This also splits the types of coordinate errors into two distinct metrics,
one for clients and one for servers, because these two should be treated
by federation admins differently. It's a much larger issue when a cache/
origin's coordinates are not known than it is when a client's coordinates
is not known.

Finally, this commit reworks the `Coordinate` struct to also keep track
of the accuracy radius and provenance information about where the
coordinate's resolution came from. These values are always tied to a
lat/long so they should be kept together.