Populate unknown IOLoad as -1 in server ads
After looking at some monitoring data with PatrickB, it looks like
it's valid/reasonable for some caches (mostly in Europe) to report
0 load if they're under utilized.

This makes the way we previously handled IO Load values problematic
because we used 0 to represent "unknown". Instead, we need a different
value for "unknown". I chose -1 because a valid IOLoad value (which
comes from the derivative of a monotonic, increasing counter) will
never be negative.