Refactor sort entrypoint, define SortAlg interface, spruce up SwapMaps, add helpers
This commit combines several veins of work that were getting harder to split
out in git history.

First, `sortServerAds`, which is called by the Director when it knows a
slice of cache/origin ads should be sorted, is modified to provide a
new entrypoint into the various sort algorithms that take the server
ads as input. In this vein of work, I define the `SortAlgorithm` interface
which supports various sort algorithms via the `.Sort()` method. Individual
implementations for each algorithm (random, distance, distanceAndLoad, adaptive)
to follow.

Second, I'm doing a bit of re-organizing, in particular moving a few types/
functions into `sort_algorithms.go`. These mostly involve functions on
`SwapMaps`, which I also spruce up with stochastic sort and descending sort
algorithms. These algorithms are a separate concept from the `SortAlgorithms`
interface because these sort `SwapMaps` and not slices of server ads. A slice
of sorted `SwapMaps` can then be used to sort the server ads.

Finally, I'm starting to touch up a few helper functions like the gated halving
multiplier. This function is switched to a continues exponential decay func
that gives the same values at the closed side of the previous gated multplier's
discontinuities.