Add weight calculation helper functions
There are two layers of functions for calculating weights -- those
with 'Fn' in their name and those without.

The functions _without_ 'Fn' in their names operate on raw inputs and
output float64 derived weights given an input.

The 'Fn'-suffixed functions wrap these other functions and handle
cases where some input or some derived weight makes the overall
calculation invalid. These functions will be called in the algorithms
so we can detect whether a weight should be used or tossed. Whenenever
a weight is tossed, the median from the set of known weights for that
weight type will be imputed. This approach has the effect of placing
unknowns in the middle of the list along a given weight axis. For example,
a server with unknown IO Load will be assigned an IO load weight such
that upon sorting along that axis it falls in the middle of the list --
half of the known servers will have a better weight and half will have
a worse weight.