Explicitly remove lr privileges for a namespace when it's write-only
Emma pointed out I hadn't completely followed my own design spec! Whenever
a namespace is managed by an Origin, it must be _explicit_ in the authfile.
What I had missed is that this means _every_ namespace must either construct
a `lr` or a `-lr` entry in the authfile.

This let me simplify the logic in `constructAuthEntry` a bit as well.