Fix bug that prevented Origins with `Logging.Level: Warn` from starting
The crux of the issue was a conversion of Pelican log levels through
logrus:
```
Warn --> (parse with Logrus) --> Warning --> apply default level on sub-systems
```

This meant that when you set `Logging.Level: Warn`, we'd parse the level
and use the parsed level to set defaults for things like `Logging.Origin.Cms`.

However, when we mapped the derived default for `param.Logging_Origin_Cms`
into xrootd config, the mapping failed because we didn't check for `warning`.

This fix is two-fold:
- first, I map "warning" back to "warn" when we're setting
defaults for the subsystems.
- second, I modified the xrootd logging code to treat "warning"
the same as "Warn"

Either of these fixes would solve the original bug, but I'm hoping that
the application of both makes a future regression that much less likely.

(cherry picked from commit 6a8d607f87869836453a485e0eff99daceb9724e)