Add Adaptive Sort knob and touch up other constants
After running our Adaptive Sort experiment for ~1 day, we've been able
to look at concrete statistics to see where there's room for improvement
without another overhaul.

The two things that stood out were large over-prioritization of load weights
(the halving threshold/factors were so small, most of the caches we observed
had load numbers that resulted in weights with 10-60 zeros at the beginning)
and that we should be more restrictive about geolocation (we're still sending
clients to caches that are a little too geographically distributed).

This commit adds a tunable knob for setting the algorithm's initial geo-truncate
value, and changes the load constants used when deriving load weights.