Fix flaky authfile test
There were previously a few issues with the test that I think stemmed
from my waffling on how to handle serializing authfile lines when two
prefixes were the same length.

Ultimately the ordering of these isn't critical to security, but a stable
sorting is really beneficial to making test assertions. While it would
be nice to preserve as much of the authline ordering as possible when we're
merging an admin-supplied authfile with the service's derived authfile,
the current data structure (a map) that holds this makes that hard to do.

Instead, this implements a stable, lexicographical sort on these values.