Point federation-based monitoring scitokens cfg at Federation issuer
In a high-availability Director setup, each Director will be signing
monitoring tokens using the federation's root hostname as the issuer.
This change makes sure Origins/Caches know to expect the new monitoring.

Fallback handling related to Director tests for older Origins/Caches
was added in #2634; that PR adjusted token signing in the Director to
first try the Federation URL and then fallback to the active Director's
own external web URL on failure.