Adjust federation discovery so we always find a federation host
I think this change mostly affects local developer setups where we
often run services with no federation root. In those cases, prior
to this change, the Issuer URL for the Director monitoring issuer
appeared blank.

This change adjusts things to that we always discover a federation
discovery URL. When none is set explicltly, we fall back to use the
Director's own hostname, since it (in theory) duplicates the same
information as a federation root.