Deprecate Debug param and map old behavior into Logging.Level
This deprecates the Debug param. To preserve the behavior of the
--debug flag, we now set Logging.Level: debug whenever the flag is
passed. Moreover, this preserves the "Debug: true" override behavior
we used to have (I actually tried to get rid of it but there are
a few technical challenges preventing me, namely that we can't
tell through viper whether some variable comes from a default or a
user -- this means I couldn't implement the logic needed to only
override Logging.Level if it's not already set).

The deprecation handling for `Debug` required a bit of special work
because we can't map the boolean value from "Debug: <bool>"
directly into `Logging.Level`, which expects a string.

Also, note that the parameters.yaml `replacedby` key is meant for direct
replacements, but as I just stated these can't be directly mapped. I still
use the `replacedby` key here because it didn't feel right to have
`replacedby: none` when it _is_ replaced, just by a param that expects a
different type. I chose instead to slightly abuse the `replacedby` key by
using the aforementioned custom logic so that our docs still point readers
toward the right replacement.