Split default log levels for clients/servers
Issue #2350 asked about cranking up the log level to 'info' for servers.
Separately, @h2zh noted that the client doesn't print warning-level
output because the log levels are too low. Since I was in the guts of
config and logging levels, I decided to additionally attempt these
two in this PR.

As for cranking log levels, I don't think we want to use info as the default for
clients, which further necessitated splitting the defaults apart. This doesn't
mean you can configure the log levels separately, though (an approach for
which the group still needs to discuss so we get it right).

There's some nuance in the defaults split because we set internal log levels before
we've constructed defaults in `config.Set{Client/Server}Defaults()`.
That meant I couldn't simply set different defaults in those functions.
Instead, I needed to duplicate where the default log levels for clients vs servers
are set both in their defaults generators and in `InitConfigInternal`.

Technically I'm changing the signature for an exported function (generally a
no-no) but so be it! I think I have a good reason for the change.
I also tried to keep the change relatively minimal and flexible by
having the `InitConfigInternal` caller pass the default log level it
wants to use. This means the caller doesn't need to know whether it's a
server or a client -- it only has to say a log level.