Always require fed discovery and provide unit tests for discoverFederationImpl
Okay, it should be _almost always_ require fed discovery, because we still
don't require it if all fed values (other than director advertis endpoints) are
set. We probably should have done this a long time ago, because a growing number
of Pelican services rely on a "federation issuer URL", but the old discovery logic
made it very easy to leave that empty.

At a basic level, the new logic is "if any fed services are missing in
config, perform federation discovery to fill them out. Keep configured
values even if we discover a different value because that's what the local
admin provided."