Fix client tests that needed to manage some aspects of discovery
These tests needed the ability to set discovery params directly to
simulate combining global config with passed Pelican URLs. The nuance
here is that we still use the mocked fed root for discovery, but we
manually manage how the client decides which config param to use for
fallback discovery. This requires unsetting some config that are set
when mocking the fed root.