Fix xroot scitokens test that was missing an issuer
This test was a bit broken because the embedded scitokens file we
compare the generated output to was entirely missing a Director/Fed-based
monitoring issuer.

Fixing it meant adding the issuer to the file, but that required jumping
through a few hoops because the MockFederationRoot function uses a dynamic
httpstest server. To handle that, I set up the mock fed root, let discovery
run and then overwrite the value that XRootD ends up pulling with something
static.