Give 'Logging.Level' a default: none label in parameters.yaml
This isn't the cleanest way to do things, but our internal stuff doesn't
currently handle a missing `Default` key. I'd argue it should if the defaults
are otherwise specified by other keys, but I don't want to pull that thread
right now because this is holding a release.

This commit restores `Default: none`, and I'm hoping docs readers will see
that the defaults are otherwise defined.