"Silence usage" for 'pelican token' commands
I did this while I was working on docs for the tool. Adding the
SilenceUsage option to cobra prevents the command from dumping all
the options/flags/usage info in the event of an error. I prefer this
because most errors the command produces don't warrant spamming people
with all this info.