Add basic user docs for working with protected data (tokens)
This documentation is meant to live at the level of "how can I as
a user get protected data", not how an extensive explanation of how
the Pelican token ecosystem works or how Origin admins can set up their
services with OAuth2 providers.

I plan to write a more extensive "how tokens in Pelican" work page
in the "Advanced" section, and this commit is a checkpoint while I
work toward that. As such, I already link to that page (which I have
sketched out locally).