Update authfile generation to work with more 'privileges'
My initial passthrough of this code assumed we'd only need to worry
about 'lr' privs, but a user who upgraded to v7.20 and was using
'u blah /foo a' in a custom authfile requested we support other privs
as well.

This commit adds all the other authfile privs except for 'k', which
the XRootD docs say is for 'lock a file' but is unused.

I also did a touch more splitting apart of structs and renamed "policy"
to "privilege" in may places -- my view is that a policy is a set
of privileges applied to a user and a path. I wanted to make that a
bit clearer, so wherever I was referring to the privilege letters
I switched to "priv".

(cherry picked from commit 56fe0cb4c9abeb47adbabe576d88327f43113f3c)