Add unit test that fails if deprecated config replacement lives in defaults yamls
Because config defaults from the defaults yamls don't work with automatic
deprecation mapping, I wanted a way to make sure any time a param is
deprecated its replacement keys won't be configured via these files.

After a quick chat with other Pelican devs, we decided the best way to
do this would be a unit test that fails if a deprecated config replacement
key is found in one of the defaults yaml files.

This test fails in those cases, and also checks an intentionally-bad
yaml string to make sure these keys are in fact caught in the
first place.

The underlying issue was that defaults from `defaults.yaml` or `osdfDefaults.yaml`
do not show up to Viper as being defaults, rather they appear as user-supplied.
For a deprecated key like `DisableHttpProxy` that's replaced by `Client.DisableHttpProxy`,
we only map the value for the deprecated key as the value for the new key if
the new key is not set by the user. This means that when replacement keys are
defined in the default yamls, we weren't doing the mapping.