Clean up deprecation handling for Xrootd.Port
There was a really hacky/funky method doing this before that relied
on defaults for `Cache.Port` and `Origin.Port` being set in the
defaults yaml. I originally missed that I broke this, but a test
failure pointed at it.

Rather than stick with the hacks, which were partially used to set
a port of "any" for xrootd's auto-port section feature when the
configured Origin/Cache port was 0, I got rid of the hack by moving
the "if Origin.Port == 0, set xrootd port to 'any'" into the Cache/
Origin templates directly.