Disable Broker for site-local caches
Right now I'm not thinking site-local caches are meant to work with
Brokers, although that may change in the future.

More importantly, since the Broker code isn't mature and there is
zero documentation for how one would test it, I have no way of knowing
whether I set up the site-local cache to work with it correctly.

For those reasons, disable the Broker feature when running in site-local
mode.